Remove message_state.c
This module was quite pointless. In the process of refactoring the
whole code base, it has now been removed. Important stuff has been
temporarily moved to core.c.

Actually this makes sense as core.c will later include only the
routines related to the main loop and synchronization, while
the various API will be moved in a separate module.

Signed-off-by: Alessandro Pellegrini <alessandro@pellegrini.tk>