Fix handling error of the reversing window.
The reversing window was being reset only in the thread_loop() function
whenever an actual reversing procedure took place. Howver this was not
ensuring that in the case the same event would being processed again as
'unsafe' the reversing window was cleared properly.
Specifically an additional unproper push %rax instruction was being added
each time the undo procedure has being called.

A reset invocation is added at the end of the execute_undoi_event() function.

Signed-off-by: Davide Cingolani<cingodvd@gmail.com>