Fix bugs in initialization structures procedure.
The wait_who vector was being initialized in the init() function with
a wrong size, specifically as an array of 'simtime_t' whereas it represents
an array of simple unsigned int.
In addition region_lock array was not being initialized properly at zero.

Has been moved the two initialization functions init() and process_init_event()
in StartSimulation() from Setup(). This is because otherwise the variable
'n_cores' was not set, though used to initialize processing time vector.