Fix wrong displacement in agent_position matrix
In 'core.c' uses agent_position matrix to retrieve the actual position
of an agent in case of MOVE events. However the displacement used was
the destination LP' id instead of the ID of the agent, actually.

Simply has been replaced the old 'receiver' field with the 'entity1'
field as offset.

Signed-off-by: tankwalker <cingodvd@gmail.com>