Zero-copy calendar queue
This commit changes the way the calendar queue manages its data
structures. In particular, the message is never copied multiple times
now.

Additionally, there was another copy of the message made by the
encapsulation layer of the calendar queue, which has been removed.

Therefore, the only malloc() call in the calendar queue is done in the
new calqueue_insert() function. Everything, then, works with pointers.

To free a message, the core must not use free() now, rather the
calqueue_free() new API.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>